requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
Flask==2.0.2
Flask-Admin==1.5.8
Flask-BasicAuth==0.2.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
peewee==3.14.8
psycopg2==2.9.2
pycparser==2.21
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2
WTForms==3.0.0
